Create README.md
In this project , we delve into the sweet world of Jelly Belly jelly beans. The purpose of this analysis is to extract meaningful insights from various facets of Jelly Belly products, ranging from the plethora of flavors to the demographics of their fan club members, the various packaging options, and the effectiveness of promotional strategies.